# docs/swagger.yaml
definitions:
  account.Account:
    properties:
      bio:
        type: string
      cover_image:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      email:
        type: string
      follower_count:
        type: integer
      following_count:
        type: integer
      full_name:
        type: string
      is_premium:
        type: boolean
      is_verified:
        type: boolean
      joined_at:
        type: string
      last_active:
        type: string
      location:
        type: string
      profile_image:
        type: string
      social_links:
        items:
          type: string
        type: array
      updated_at:
        type: string
      updated_by:
        type: string
      username:
        type: string
      website:
        type: string
    type: object
  account.Login:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  account.LoginResponse:
    properties:
      access_Token:
        type: string
      id:
        type: string
      refresh_Token:
        type: string
    type: object
  account.SignUp:
    properties:
      email:
        type: string
      id:
        type: string
      password:
        type: string
    type: object
  account.SignUpRes:
    properties:
      access_Token:
        type: string
      id:
        type: string
      refresh_Token:
        type: string
    type: object
  account.UpdateAccount:
    properties:
      bio:
        type: string
      cover_image:
        type: string
      email:
        type: string
      full_name:
        type: string
      location:
        type: string
      profile_image:
        type: string
      social_links:
        items:
          type: string
        type: array
      username:
        type: string
      website:
        type: string
    type: object
  post.BlogPost:
    properties:
      author:
        $ref: '#/definitions/post.User'
      author_id:
        type: string
      content:
        items:
          additionalProperties: true
          type: object
        type: array
      cover_image:
        type: string
      id:
        type: string
      published_at:
        type: string
      read_time:
        type: integer
      status:
        type: integer
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      visibility:
        type: string
    type: object
  post.UpdateBlogPost:
    properties:
      author:
        $ref: '#/definitions/post.User'
      author_id:
        type: string
      content:
        items:
          additionalProperties: true
          type: object
        type: array
      cover_image:
        type: string
      published_at:
        type: string
      read_time:
        type: integer
      status:
        type: integer
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      visibility:
        type: string
    type: object
  post.User:
    properties:
      email:
        type: string
      full_name:
        type: string
      profile_image:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/account:
    get:
      consumes:
      - application/json
      description: Get Account Details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/account.Account'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get Account Details
      tags:
      - Account
  /api/v1/account/edit:
    put:
      consumes:
      - application/json
      description: Update Account Details
      parameters:
      - description: Update Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/account.UpdateAccount'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - Bearer: []
      summary: Update Account Details
      tags:
      - Account
  /api/v1/account/login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: SignUp Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/account.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/account.LoginResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Login
      tags:
      - Account
  /api/v1/account/signup:
    post:
      consumes:
      - application/json
      description: SignUp
      parameters:
      - description: SignUp Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/account.SignUp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/account.SignUpRes'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: SignUp
      tags:
      - Account
  /api/v1/blog/post:
    get:
      consumes:
      - application/json
      description: Get All Post
      produces:
      - application/json
      responses:
        "200":
          description: sucess
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get All Post
      tags:
      - Post
    post:
      consumes:
      - application/json
      description: Create Post
      parameters:
      - description: Create Post Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/post.BlogPost'
      produces:
      - application/json
      responses:
        "200":
          description: sucess
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - Bearer: []
      summary: Create Post
      tags:
      - Post
  /api/v1/blog/post/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Post
      parameters:
      - description: Post Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: sucess
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - Bearer: []
      summary: Delete Post
      tags:
      - Post
    get:
      consumes:
      - application/json
      description: Get Post By Id
      parameters:
      - description: Post Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: sucess
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get Post By Id
      tags:
      - Post
    put:
      consumes:
      - application/json
      description: Update Post
      parameters:
      - description: Create Post Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/post.UpdateBlogPost'
      - description: Post Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: sucess
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - Bearer: []
      summary: Update Post
      tags:
      - Post
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"